+++ src/lib_tests.rs
@@ -1244,6 +1244,88 @@ fn shardbyshard_encode_correctly() {
     }
 }
 
+quickcheck! {
+    fn qc_shardbyshard_encode_same_as_encode(data   : usize,
+                                             parity : usize,
+                                             size   : usize) -> bool {
+        let data   = 1 + data % 256;
+        let mut parity = 1 + parity % 256;
+        if data + parity > 256 {
+            parity -= data + parity - 256;
+        }
+
+        let size = 1 + size % 1_000_000;
+
+        let r = ReedSolomon::new(data, parity).unwrap();
+        let mut sbs = ShardByShard::new(&r);
+
+        let mut expect = make_random_shards!(size, data + parity);
+        let mut shards = expect.clone();
+
+        {
+            let mut refs =
+                convert_2D_slices!(expect =>to_mut_vec &mut [u8]);
+
+            r.encode(&mut refs).unwrap();
+        }
+
+        let expect = expect;
+
+        {
+            let mut slice_refs =
+                convert_2D_slices!(shards      =>to_mut_vec &mut [u8]);
+
+            for i in 0..data {
+                assert_eq!(i, sbs.cur_input_index());
+
+                sbs.encode(&mut slice_refs).unwrap();
+            }
+        }
+
+        let shards = shards;
+
+        expect == shards
+            && sbs.parity_ready()
+            && sbs.cur_input_index() == data
+            && { sbs.reset().unwrap(); !sbs.parity_ready() && sbs.cur_input_index() == 0 }
+    }
+
+    fn qc_shardbyshard_encode_same_as_encode_shards(data   : usize,
+                                                    parity : usize,
+                                                    size   : usize) -> bool {
+        let data   = 1 + data % 256;
+        let mut parity = 1 + parity % 256;
+        if data + parity > 256 {
+            parity -= data + parity - 256;
+        }
+
+        let size = 1 + size % 1_000_000;
+
+        let r = ReedSolomon::new(data, parity).unwrap();
+        let mut sbs = ShardByShard::new(&r);
+
+        let mut expect = make_random_shards!(size, data + parity);
+        let mut shards = expect.clone();
+
+        r.encode_shards(&mut expect).unwrap();
+
+        let expect = expect;
+
+        for i in 0..data {
+            assert_eq!(i, sbs.cur_input_index());
+
+            sbs.encode_shard(&mut shards).unwrap();
+        }
+
+        let shards = shards;
+
+        expect == shards
+            && sbs.parity_ready()
+            && sbs.cur_input_index() == data
+            && { sbs.reset().unwrap(); !sbs.parity_ready() && sbs.cur_input_index() == 0 }
+    }
+}
+
 #[test]
 fn shardbyshard_encode_sep_correctly() {
     {
@@ -1320,6 +1402,103 @@ fn shardbyshard_encode_sep_correctly() {
         sbs.reset_force();
 
         assert_eq!(0, sbs.cur_input_index());
+    }
+}
+
+quickcheck! {
+    fn qc_shardbyshard_encode_sep_same_as_encode(data   : usize,
+                                                 parity : usize,
+                                                 size   : usize) -> bool {
+        let data   = 1 + data % 256;
+        let mut parity = 1 + parity % 256;
+        if data + parity > 256 {
+            parity -= data + parity - 256;
+        }
+
+        let size = 1 + size % 1_000_000;
+
+        let r = ReedSolomon::new(data, parity).unwrap();
+        let mut sbs = ShardByShard::new(&r);
+
+        let mut expect = make_random_shards!(size, data + parity);
+        let mut shards = expect.clone();
+
+        {
+            let (data_shards, parity_shards) =
+                expect.split_at_mut(data);
+
+            let data_refs =
+                convert_2D_slices!(data_shards   =>to_mut_vec &[u8]);
+            let mut parity_refs =
+                convert_2D_slices!(parity_shards =>to_mut_vec &mut [u8]);
+
+            r.encode_sep(&data_refs, &mut parity_refs).unwrap();
+        }
+
+        let expect = expect;
+
+        {
+            let (data_shards, parity_shards) =
+                shards.split_at_mut(data);
+            let data_refs =
+                convert_2D_slices!(data_shards   =>to_mut_vec &[u8]);
+            let mut parity_refs =
+                convert_2D_slices!(parity_shards =>to_mut_vec &mut [u8]);
+
+            for i in 0..data {
+                assert_eq!(i, sbs.cur_input_index());
+
+                sbs.encode_sep(&data_refs, &mut parity_refs).unwrap();
+            }
+        }
+
+        expect == shards
+            && sbs.parity_ready()
+            && sbs.cur_input_index() == data
+            && { sbs.reset().unwrap(); !sbs.parity_ready() && sbs.cur_input_index() == 0 }
+    }
+
+    fn qc_shardbyshard_encode_sep_same_as_encode_shards(data   : usize,
+                                                        parity : usize,
+                                                        size   : usize) -> bool {
+        let data   = 1 + data % 256;
+        let mut parity = 1 + parity % 256;
+        if data + parity > 256 {
+            parity -= data + parity - 256;
+        }
+
+        let size = 1 + size % 1_000_000;
+
+        let r = ReedSolomon::new(data, parity).unwrap();
+        let mut sbs = ShardByShard::new(&r);
+
+        let mut expect = make_random_shards!(size, data + parity);
+        let mut shards = expect.clone();
+
+        {
+            let (data_shards, parity_shards) =
+                expect.split_at_mut(data);
+
+            r.encode_shards_sep(data_shards, parity_shards).unwrap();
+        }
+
+        let expect = expect;
+
+        {
+            let (data_shards, parity_shards) =
+                shards.split_at_mut(data);
+
+            for i in 0..data {
+                assert_eq!(i, sbs.cur_input_index());
+
+                sbs.encode_shard_sep(data_shards, parity_shards).unwrap();
+            }
+        }
+
+        expect == shards
+            && sbs.parity_ready()
+            && sbs.cur_input_index() == data
+            && { sbs.reset().unwrap(); !sbs.parity_ready() && sbs.cur_input_index() == 0 }
     }
 }