+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 - Upgraded InversionTree QuickCheck test
   - No code changes as no bugs were found
 - Upgraded test suite for main codec methods(e.g. encode, reconstruct)
-  - A lot heavy QuickCheck tests were added
+  - A lot of heavy QuickCheck tests were added
+  - No code changes as no bugs were found
+- Upgraded test suite for ShardByShard methods
+  - A lot of heavy QuickCheck tests were added
   - No code changes as no bugs were found
 
 ## 3.0.3